+++ supabase/migrations/20260506112311_add_posts_tags_table.sql
@@ -1,0 +1,7 @@
+CREATE TABLE public.posts_tags(
+    id uuid PRIMARY KEY DEFAULT gen_random_uuid(),
+    tag_id uuid NOT NULL REFERENCES public.post_tags(id) ON DELETE CASCADE
+);
+
+ALTER TABLE public.posts_tags
+ENABLE ROW LEVEL SECURITY;